get_status: Process input in batches
When bugs.debian.org was upgraded to Jessie, resources limits were set
in the Apache configuration.  This means that process of get_status
requests for a large number of bugs will cause the process to get
killed, due to exceeding the resource limits.

Processing the input to get_status in batches avoids hitting the
resource limits.