fixed ssl validation on python 3.4.3 and httlib2
to support python 3.4.3, when using httplib2 as pysimplesoap transport we must
work around a bug in httplib2, which uses http.client.HTTPSConnection with
check_hostname=True, but with an empty ssl context that prevents the
certificate verification. Passing `cacert` to httplib2 through pysimplesoap
permits to create a valid ssl context.
With python >= 3.4.4 the empty context means DO validation, so we don't
need the workaround there.